Рефакторинг существующего кода приложения с использованием MVVM
