+++ app/src/main/java/com/ururu2909/firstapp/ContactDetailsFragment.java
@@ -16,6 +16,10 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.ururu2909.firstapp.viewmodel.ContactsViewModel;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -23,7 +27,6 @@ import java.util.Calendar;
 import java.util.Locale;
 
 public class ContactDetailsFragment extends Fragment implements CompoundButton.OnCheckedChangeListener {
-    private ContactsService mService;
     private TextView contactName;
     private TextView contactPhoneNumber1;
     private TextView contactPhoneNumber2;
@@ -31,17 +34,6 @@ public class ContactDetailsFragment extends Fragment implements CompoundButton.O
     private TextView contactEmail2;
     private TextView contactBirthDate;
     private AlarmManager alarmManager;
-
-    interface ResultListener {
-        void onComplete(Contact result);
-    }
-    @Override
-    public void onAttach(@NonNull Context context) {
-        super.onAttach(context);
-        if (context instanceof ServiceProvider){
-            this.mService = ((ServiceProvider) context).getService();
-        }
-    }
 
     static ContactDetailsFragment newInstance(int index, String id) {
         ContactDetailsFragment f = new ContactDetailsFragment();
@@ -59,6 +51,7 @@ public class ContactDetailsFragment extends Fragment implements CompoundButton.O
         View view = inflater.inflate(R.layout.fragment_contact_details, container, false);
         int index = this.getArguments().getInt("index");
         String id = this.getArguments().getString("id");
+        ContactsViewModel model = new ViewModelProvider(requireActivity()).get(ContactsViewModel.class);
 
         contactName = view.findViewById(R.id.contactDetailsName);
         contactPhoneNumber1 = view.findViewById(R.id.contactDetailsPhoneNumber);
@@ -82,7 +75,28 @@ public class ContactDetailsFragment extends Fragment implements CompoundButton.O
             }
             birthdayNotifySwitch.setOnCheckedChangeListener(this);
         }
-        mService.getContact(callback, id);
+        model.getContact(id).observe(getViewLifecycleOwner(), new Observer<Contact>() {
+            @Override
+            public void onChanged(final Contact contact) {
+                if (contactName != null && contactPhoneNumber1 != null && contactPhoneNumber2 !=null
+                        && contactEmail1 !=null && contactEmail2 !=null && contactBirthDate !=null){
+                    contactName.post(new Runnable() {
+                        @Override
+                        public void run() {
+                            if (contactName != null && contactPhoneNumber1 != null && contactPhoneNumber2 !=null
+                                    && contactEmail1 !=null && contactEmail2 !=null && contactBirthDate !=null) {
+                                contactName.setText(contact.getName());
+                                contactPhoneNumber1.setText(contact.getPhoneNumber1());
+                                contactPhoneNumber2.setText(contact.getPhoneNumber2());
+                                contactEmail1.setText(contact.getEmail1());
+                                contactEmail2.setText(contact.getEmail2());
+                                contactBirthDate.setText(contact.getBirthDate());
+                            }
+                        }
+                    });
+                }
+            }
+        });
         return view;
     }
 
@@ -97,30 +111,6 @@ public class ContactDetailsFragment extends Fragment implements CompoundButton.O
         contactEmail2 = null;
         contactBirthDate = null;
     }
-
-    private ContactDetailsFragment.ResultListener callback = new ContactDetailsFragment.ResultListener() {
-        @Override
-        public void onComplete(Contact result) {
-            final Contact contact = result;
-            if (contactName != null && contactPhoneNumber1 != null && contactPhoneNumber2 !=null
-                    && contactEmail1 !=null && contactEmail2 !=null && contactBirthDate !=null){
-                contactName.post(new Runnable() {
-                    @Override
-                    public void run() {
-                        if (contactName != null && contactPhoneNumber1 != null && contactPhoneNumber2 !=null
-                                && contactEmail1 !=null && contactEmail2 !=null && contactBirthDate !=null) {
-                            contactName.setText(contact.getName());
-                            contactPhoneNumber1.setText(contact.getPhoneNumber1());
-                            contactPhoneNumber2.setText(contact.getPhoneNumber2());
-                            contactEmail1.setText(contact.getEmail1());
-                            contactEmail2.setText(contact.getEmail2());
-                            contactBirthDate.setText(contact.getBirthDate());
-                        }
-                    }
-                });
-            }
-        }
-    };
 
     @Override
     public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
