+++ app/src/main/java/com/ururu2909/firstapp/MainActivity.java
@@ -11,22 +11,12 @@ import androidx.fragment.app.FragmentTransaction;
 import android.Manifest;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
-import android.content.ComponentName;
-import android.content.Context;
-import android.content.Intent;
-import android.content.ServiceConnection;
 import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.IBinder;
 import android.widget.Toast;
 
-import com.ururu2909.firstapp.ContactsService.MyBinder;
-
-public class MainActivity extends AppCompatActivity implements ServiceProvider, ActivityCompat.OnRequestPermissionsResultCallback {
-    ContactsService mService;
-    boolean mBound = false;
-    boolean createdFirstTime;
+public class MainActivity extends AppCompatActivity implements ActivityCompat.OnRequestPermissionsResultCallback {
     public static final int PERMISSION_REQUEST_CODE = 10;
 
     @RequiresApi(api = Build.VERSION_CODES.M)
@@ -35,37 +25,14 @@ public class MainActivity extends AppCompatActivity implements ServiceProvider, 
         super.onCreate(savedInstanceState);
         createNotificationChannel();
         setContentView(R.layout.activity_main);
-        createdFirstTime = savedInstanceState == null;
         if (ContextCompat.checkSelfPermission(this, android.Manifest.permission.READ_CONTACTS) == PackageManager.PERMISSION_GRANTED) {
-            Intent intent = new Intent(MainActivity.this, ContactsService.class);
-            bindService(intent, mConnection, Context.BIND_AUTO_CREATE);
+            if (savedInstanceState == null){
+                addFragment();
+            }
         } else {
             requestPermissions(new String[]{Manifest.permission.READ_CONTACTS}, PERMISSION_REQUEST_CODE);
         }
     }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        unbindService(mConnection);
-    }
-
-    private ServiceConnection mConnection = new ServiceConnection() {
-        @Override
-        public void onServiceConnected(ComponentName name, IBinder service) {
-            MyBinder binder = (MyBinder) service;
-            mService = binder.getService();
-            mBound = true;
-            if (createdFirstTime){
-                addFragment();
-            }
-        }
-
-        @Override
-        public void onServiceDisconnected(ComponentName name) {
-            mBound = false;
-        }
-    };
 
     void addFragment(){
         int index = getIntent().getIntExtra("contactIndex", -1);
@@ -82,11 +49,6 @@ public class MainActivity extends AppCompatActivity implements ServiceProvider, 
             ft.add(R.id.container, fragment);
             ft.commit();
         }
-    }
-
-    @Override
-    public ContactsService getService() {
-        return mService;
     }
 
     private void createNotificationChannel() {
@@ -106,10 +68,8 @@ public class MainActivity extends AppCompatActivity implements ServiceProvider, 
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
         switch (requestCode) {
             case PERMISSION_REQUEST_CODE:
-                if (grantResults.length > 0 &&
-                        grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    Intent intent = new Intent(MainActivity.this, ContactsService.class);
-                    bindService(intent, mConnection, Context.BIND_AUTO_CREATE);
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                    addFragment();
                 } else {
                     Toast.makeText(this, R.string.need_permission_msg, Toast.LENGTH_LONG).show();
                 }
