--- app/src/main/java/com/ururu2909/firstapp/ServiceProvider.java
@@ -1,5 +1,0 @@
-package com.ururu2909.firstapp;
-
-public interface ServiceProvider {
-    ContactsService getService();
-}
